+++ tests/testthat/test-path.R
@@ -42,9 +42,18 @@ describe("path", {
   })
 
   it("errors on paths which are too long", {
-    expect_error(fs::path(paste(rep("a", 100000), collapse = "")), "less than PATH_MAX")
+    expect_error(path(paste(rep("a", 100000), collapse = "")), "less than PATH_MAX")
 
-    expect_error(do.call(fs::path, as.list(rep("a", 100000))), "less than PATH_MAX")
+    expect_error(do.call(path, as.list(rep("a", 100000))), "less than PATH_MAX")
+  })
+
+  it("follows recycling rules", {
+    expect_equal(path("foo", character()), character())
+    expect_equal(path("foo", "bar"), "foo/bar")
+    expect_equal(path("foo", c("bar", "baz")), c("foo/bar", "foo/baz"))
+    expect_equal(path(c("foo", "qux"), c("bar", "baz")), c("foo/bar", "qux/baz"))
+
+    expect_error(path(c("foo", "qux", "foo2"), c("bar", "baz")), "arguments must have consistent lengths", class = "invalid_argument")
   })
 })